+++ app/api/auth/[...all]/route.ts
@@ -1,7 +1,4 @@
-// app/api/auth/[...all]/route.ts
-import { auth } from "@/lib/auth/config";
+import { auth } from "@/lib/auth";
 import { toNextJsHandler } from "better-auth/next-js";
 
-const handler = toNextJsHandler(auth);
-
-export { handler as GET, handler as POST };
+export const {  GET, POST } = toNextJsHandler(auth.handler);
