+++ lib/auth.ts
@@ -1,8 +1,8 @@
 // lib/auth/config.ts
 import { betterAuth } from "better-auth";
 import { drizzleAdapter } from "better-auth/adapters/drizzle";
-import { db } from "../db";
-import * as schema from "../db/schema";
+import { db } from "@/lib/db";
+import * as schema from "@/lib/db/schema";
 
 export const auth = betterAuth({
   database: drizzleAdapter(db, {
@@ -27,13 +27,14 @@ export const auth = betterAuth({
       clientSecret: process.env.GITHUB_CLIENT_SECRET as string,
     },
   },
-  session: {
-    expiresIn: 60 * 60 * 24 * 7, // 7 days
-    updateAge: 60 * 60 * 24, // 24 hours
-  },
-  advanced: {
-    generateId: () => crypto.randomUUID(),
-  },
+  // session: {
+  //   expiresIn: 60 * 60 * 24 * 7, // 7 days
+  //   updateAge: 60 * 60 * 24, // 24 hours
+  // },
+  // advanced: {
+  //   generateId: () => crypto.randomUUID(),
+  // },
+  baseURL: process.env.NEXT_PUBLIC_BASE_URL
 });
 
 // Export types for client-side usage
