+++ middleware.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { auth } from "./lib/auth/config";
+import { auth } from "./lib/auth";
 
 export default async function middleware(request: NextRequest) {
   const { pathname } = request.nextUrl;
